patent_backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# CORS (Next.js 연동용)
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# Data Processing
pandas==2.2.3
numpy==2.2.1

# AI/ML Libraries
openai==1.58.1
# langchain==0.3.13
# langchain-openai==0.2.14

# Vector Search (선택)
# pgvector==0.2.4
# pinecone-client==3.0.0

# File Upload & Processing
Pillow==10.4.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.2

# API Documentation
drf-spectacular==0.27.1

# Async & Background Tasks
celery==5.3.6
redis==5.2.1
django-celery-beat==2.6.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring (선택)
# sentry-sdk==1.40.0
